+++ types.go
@@ -71,9 +71,9 @@ type Label struct {
 }
 
 type RunnerGroup struct {
-	ID        int    `json:"id"`
+	ID        uint64 `json:"id"`
 	Name      string `json:"name"`
-	Size      int64  `json:"size"`
+	Size      uint64 `json:"size"`
 	IsDefault bool   `json:"isDefaultGroup"`
 }
 
@@ -83,9 +83,9 @@ type RunnerGroupList struct {
 }
 
 type RunnerScaleSet struct {
-	ID                 int                      `json:"id,omitempty"`
+	ID                 uint64                   `json:"id,omitempty"`
 	Name               string                   `json:"name,omitempty"`
-	RunnerGroupID      int                      `json:"runnerGroupId,omitempty"`
+	RunnerGroupID      uint64                   `json:"runnerGroupId,omitempty"`
 	RunnerGroupName    string                   `json:"runnerGroupName,omitempty"`
 	Labels             []Label                  `json:"labels,omitempty"`
 	RunnerSetting      RunnerSetting            `json:"RunnerSetting,omitempty"`
@@ -100,14 +100,14 @@ type RunnerScaleSetJitRunnerSetting struct {
 }
 
 type RunnerScaleSetMessage struct {
-	MessageID   int64                    `json:"messageId"`
+	MessageID   uint64                   `json:"messageId"`
 	MessageType string                   `json:"messageType"`
 	Body        string                   `json:"body"`
 	Statistics  *RunnerScaleSetStatistic `json:"statistics"`
 }
 
 type runnerScaleSetsResponse struct {
-	Count           int              `json:"count"`
+	Count           uint64           `json:"count"`
 	RunnerScaleSets []RunnerScaleSet `json:"value"`
 }
 
@@ -121,13 +121,13 @@ type RunnerScaleSetSession struct {
 }
 
 type RunnerScaleSetStatistic struct {
-	TotalAvailableJobs     int `json:"totalAvailableJobs"`
-	TotalAcquiredJobs      int `json:"totalAcquiredJobs"`
-	TotalAssignedJobs      int `json:"totalAssignedJobs"`
-	TotalRunningJobs       int `json:"totalRunningJobs"`
-	TotalRegisteredRunners int `json:"totalRegisteredRunners"`
-	TotalBusyRunners       int `json:"totalBusyRunners"`
-	TotalIdleRunners       int `json:"totalIdleRunners"`
+	TotalAvailableJobs     uint64 `json:"totalAvailableJobs"`
+	TotalAcquiredJobs      uint64 `json:"totalAcquiredJobs"`
+	TotalAssignedJobs      uint64 `json:"totalAssignedJobs"`
+	TotalRunningJobs       uint64 `json:"totalRunningJobs"`
+	TotalRegisteredRunners uint64 `json:"totalRegisteredRunners"`
+	TotalBusyRunners       uint64 `json:"totalBusyRunners"`
+	TotalIdleRunners       uint64 `json:"totalIdleRunners"`
 }
 
 type RunnerSetting struct {
